requirements.txt
pandas
numpy
scikit-learn
xgboost
google-cloud-storage
google-cloud-aiplatform
cloudml-hypertune
